requirements.txt
streamlit==1.29.0
torch>=2.0.0
transformers>=4.35.0
qdrant-client>=1.6.0
numpy>=1.24.0
Pillow>=10.0.0
python-dotenv>=1.0.0
requests>=2.31.0
tqdm>=4.66.0
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-core>=0.1.0
langchain-groq>=0.0.3
